Finished last changes wuhu

+++ src/main/java/dk/kea/calendue/model/Project.java
@@ -10,11 +10,11 @@ public class Project
     private int project_hours;
     private String project_status;
     private int project_days;
-    private int project_hours_per_day;
+    private double project_hours_per_day;
     private int users_assigned;
     private int project_hours_scheduled;
 
-    public Project(int project_id, String project_name, String project_description, String project_start, String project_deadline, int project_hours, String project_status, int project_days, int project_hours_per_day, int users_assigned, int project_hours_scheduled)
+    public Project(int project_id, String project_name, String project_description, String project_start, String project_deadline, int project_hours, String project_status, int project_days, double project_hours_per_day, int users_assigned, int project_hours_scheduled)
     {
         this.project_id = project_id;
         this.project_name = project_name;
@@ -107,12 +107,12 @@ public class Project
         this.project_days = project_days;
     }
 
-    public int getProject_hours_per_day()
+    public double getProject_hours_per_day()
     {
         return project_hours_per_day;
     }
 
-    public void setProject_hours_per_day(int project_hours_per_day)
+    public void setProject_hours_per_day(double project_hours_per_day)
     {
         this.project_hours_per_day = project_hours_per_day;
     }
